Making it cloud agnostic
