+++ qds_sdk/commands.py
@@ -22,14 +22,8 @@ import re
 import pipes
 import os
 import json
-import uuid
 
 log = logging.getLogger("qds_commands")
-
-# Pattern matcher for s3 path
-_URI_RE = re.compile(r's3://([^/]+)/?(.*)')
-Oracle_URI_RE = re.compile(r'oraclebmc://([^@]+)([^/]+)/?(.*)')
-Azure_URI_RE = re.compile(r'wasb://([^@]+)([^/]+)/?(.*)')
 
 class Command(Resource):
 
@@ -237,67 +231,8 @@ class Command(Resource):
         else:
             if fetch:
                 storage_credentials = conn.get(Account.credentials_rest_entity_path)
-                if Qubole.cloud == 'aws':
-                    boto_conn = boto.connect_s3(aws_access_key_id=storage_credentials['storage_access_key'],
-                                                aws_secret_access_key=storage_credentials['storage_secret_key'],
-                                                security_token=storage_credentials['session_token'])
-
-                    log.info("Starting download from result locations: [%s]" % ",".join(r['result_location']))
-                    # fetch latest value of num_result_dir
-                    num_result_dir = Command.find(self.id).num_result_dir
-
-                    for s3_path in r['result_location']:
-
-                        # If column/header names are not able to fetch then use include header as true
-                        if include_header.lower() == "true" and qlog is not None:
-                            write_headers(qlog, fp)
-
-                        # In Python 3,
-                        # If the delim is None, fp should be in binary mode because
-                        # boto expects it to be.
-                        # If the delim is not None, then both text and binary modes
-                        # work.
-
-                        _download_to_local(boto_conn, s3_path, fp, num_result_dir, delim=delim,
-                                           skip_data_avail_check=isinstance(self, PrestoCommand))
-
-                elif Qubole.cloud == 'oracle_bmc':
-                    import oraclebmc
-                    private_key = storage_credentials.get('storage_config').get('api_private_rsa_key')
-                    private_key_file_path = "/tmp/" + uuid.uuid4().hex
-                    with open(private_key_file_path, 'w+') as private_key_file:
-                        private_key_file.write(private_key)
-                    os.chmod(private_key_file_path, 0o744)
-                    oracle_config = {"user" : storage_credentials.get('storage_config').get('user_id'),
-                                     "key_file" : private_key_file_path,
-                                     "fingerprint" : storage_credentials.get('storage_config').get('key_finger_print'),
-                                     "tenancy" : storage_credentials.get('storage_config').get('tenant_id'),
-                                     "region": "us-phoenix-1",  # just now added us-phoenix-1 , have to change api to get this param
-                                     "pass_phrase": None,
-                                     "additional_user_agent": "",
-                                     "log_requests": False
-                                     }
-                    object_storage = oraclebmc.object_storage.ObjectStorageClient(oracle_config)
-                    log.info("Starting download from result locations: [%s]" % ",".join(r['result_location']))
-                    num_result_dir = Command.find(self.id).num_result_dir
-                    for oracle_path in r['result_location']:
-                        _download_to_local_oracle(object_storage, oracle_path, fp, delim=delim,
-                                                  skip_data_avail_check=isinstance(self, PrestoCommand))
-
-                elif Qubole.cloud == 'azure':
-                    from azure.storage.blob import BlockBlobService
-                    account_name = storage_credentials.get('storage_config').get('account_name')
-                    account_key = storage_credentials.get('storage_config').get('access_key')
-                    block_blob_service = BlockBlobService(account_name=account_name, account_key=account_key)
-                    log.info("Starting download from result locations: [%s]" % ",".join(r['result_location']))
-                    num_result_dir = Command.find(self.id).num_result_dir
-                    for azure_path in r['result_location']:
-                        _download_to_local_azure(block_blob_service, azure_path, fp, delim=delim,
-                                                  skip_data_avail_check=isinstance(self, PrestoCommand))
-
-                else:
-                    log.info("Large results download is not supported for %s cloud" % Qubole.cloud)
-                    return
+                cloud = Qubole.get_cloud()
+                cloud.get_results(fp, storage_credentials, r, delim, qlog, include_header)
             else:
                 fp.write(",".join(r['result_location']))
 
@@ -1278,271 +1213,3 @@ class DbTapQueryCommand(Command):
         v = vars(options)
         v["command_type"] = "DbTapQueryCommand"
         return v
-
-def _read_iteratively(key_instance, fp, delim):
-    key_instance.open_read()
-    while True:
-        try:
-            # Default buffer size is 8192 bytes
-            data = next(key_instance)
-            if sys.version_info < (3, 0, 0):
-                fp.write(str(data).replace(chr(1), delim))
-            else:
-                import io
-                if isinstance(fp, io.TextIOBase):
-                    fp.buffer.write(data.decode('utf-8').replace(chr(1), delim).encode('utf8'))
-                elif isinstance(fp, io.BufferedIOBase) or isinstance(fp, io.RawIOBase):
-                    fp.write(data.decode('utf8').replace(chr(1), delim).encode('utf8'))
-                else:
-                    # Can this happen? Don't know what's the right thing to do in this case.
-                    pass
-        except StopIteration:
-            # Stream closes itself when the exception is raised
-            return
-
-def write_headers(qlog,fp):
-    col_names = []
-    qlog = json.loads(qlog)
-    if qlog["QBOL-QUERY-SCHEMA"] is not None:
-        qlog_hash = qlog["QBOL-QUERY-SCHEMA"]["-1"] if qlog["QBOL-QUERY-SCHEMA"]["-1"] is not None else qlog["QBOL-QUERY-SCHEMA"][qlog["QBOL-QUERY-SCHEMA"].keys[0]]
-
-        for qlog_item in qlog_hash:
-            col_names.append(qlog_item["ColumnName"])
-
-        col_names = "\t".join(col_names)
-        col_names += "\n"
-
-    fp.write(col_names)
-
-
-def _download_to_local(boto_conn, s3_path, fp, num_result_dir, delim=None, skip_data_avail_check=False):
-    '''
-    Downloads the contents of all objects in s3_path into fp
-
-    Args:
-        `boto_conn`: S3 connection object
-
-        `s3_path`: S3 path to be downloaded
-
-        `fp`: The file object where data is to be downloaded
-    '''
-    #Progress bar to display download progress
-    def _callback(downloaded, total):
-        '''
-        Call function for upload.
-
-        `downloaded`: File size already downloaded (int)
-
-        `total`: Total file size to be downloaded (int)
-        '''
-        if (total is 0) or (downloaded == total):
-            return
-        progress = downloaded*100/total
-        sys.stderr.write('\r[{0}] {1}%'.format('#'*progress, progress))
-        sys.stderr.flush()
-
-    def _is_complete_data_available(bucket_paths, num_result_dir):
-        if num_result_dir == -1:
-            return True
-        unique_paths = set()
-        files = {}
-        for one_path in bucket_paths:
-            name = one_path.name.replace(key_prefix, "", 1)
-            if name.startswith('_tmp.'):
-                continue
-            path = name.split("/")
-            dir = path[0].replace("_$folder$", "", 1)
-            unique_paths.add(dir)
-            if len(path) > 1:
-                file = int(path[1])
-                if dir not in files:
-                    files[dir] = []
-                files[dir].append(file)
-        if len(unique_paths) < num_result_dir:
-            return False
-        for k in files:
-            v = files.get(k)
-            if len(v) > 0 and max(v) + 1 > len(v):
-                return False
-        return True
-
-    m = _URI_RE.match(s3_path)
-    bucket_name = m.group(1)
-    bucket = boto_conn.get_bucket(bucket_name)
-    retries = 6
-    if s3_path.endswith('/') is False:
-        #It is a file
-        key_name = m.group(2)
-        key_instance = bucket.get_key(key_name)
-        while key_instance is None and retries > 0:
-            retries = retries - 1
-            log.info("Results file is not available on s3. Retry: " + str(6-retries))
-            time.sleep(10)
-            key_instance = bucket.get_key(key_name)
-        if key_instance is None:
-            raise Exception("Results file not available on s3 yet. This can be because of s3 eventual consistency issues.")
-        log.info("Downloading file from %s" % s3_path)
-        if delim is None:
-            try:
-                key_instance.get_contents_to_file(fp)  # cb=_callback
-            except boto.exception.S3ResponseError as e:
-                if (e.status == 403):
-                    # SDK-191, boto gives an error while fetching the objects using versions which happens by default
-                    # in the get_contents_to_file() api. So attempt one without specifying version.
-                    log.warn("Access denied while fetching the s3 object. Retrying without specifying the version....")
-                    key_instance.open()
-                    fp.write(key_instance.read())
-                    key_instance.close()
-                else:
-                    raise
-        else:
-            # Get contents as string. Replace parameters and write to file.
-            _read_iteratively(key_instance, fp, delim=delim)
-
-    else:
-        #It is a folder
-        key_prefix = m.group(2)
-        bucket_paths = bucket.list(key_prefix)
-        if not skip_data_avail_check:
-            complete_data_available = _is_complete_data_available(bucket_paths, num_result_dir)
-            while complete_data_available is False and retries > 0:
-                retries = retries - 1
-                log.info("Results dir is not available on s3. Retry: " + str(6-retries))
-                time.sleep(10)
-                complete_data_available = _is_complete_data_available(bucket_paths, num_result_dir)
-            if complete_data_available is False:
-                raise Exception("Results file not available on s3 yet. This can be because of s3 eventual consistency issues.")
-
-        for one_path in bucket_paths:
-            name = one_path.name
-
-            # Eliminate _tmp_ files which ends with $folder$
-            if name.endswith('$folder$'):
-                continue
-
-            log.info("Downloading file from %s" % name)
-            if delim is None:
-                one_path.get_contents_to_file(fp)  # cb=_callback
-            else:
-                _read_iteratively(one_path, fp, delim=delim)
-
-
-def _download_to_local_oracle(storage_conn, oracle_path, fp, delim=None,skip_data_avail_check=False):
-
-    def get_contents_to_file_using_delim():
-        if sys.version_info < (3, 0, 0):
-            fp.write(str(content).replace(chr(1), delim))
-        else:
-            import io
-            if isinstance(fp, io.TextIOBase):
-                fp.buffer.write(content.decode('utf-8').replace(chr(1), delim).encode('utf8'))
-            elif isinstance(fp, io.BufferedIOBase) or isinstance(fp, io.RawIOBase):
-                fp.write(content.decode('utf8').replace(chr(1), delim).encode('utf8'))
-            else:
-                # Can this happen? Don't know what's the right thing to do in this case.
-                pass
-
-    def get_contents_to_file_oracle():
-        if sys.version_info < (3, 0, 0):
-            fp.write(content.encode('utf8'))
-        else:
-            import io
-            if isinstance(fp, io.TextIOBase):
-                fp.buffer.write(content.encode('utf8'))
-            elif isinstance(fp, io.BufferedIOBase) or isinstance(fp, io.RawIOBase):
-                fp.write(content.encode('utf8'))
-            else:
-                # Can this happen? Don't know what's the right thing to do in this case.
-                pass
-
-
-    try:
-        m = Oracle_URI_RE.match(oracle_path)
-        bucket_name = m.group(1)
-        object_name = m.group(3)
-        namespace_name = storage_conn.get_namespace().data
-        if oracle_path.endswith('/') is False:
-
-            response = storage_conn.get_object(namespace_name, bucket_name, object_name)
-            content = response.data.content if response.data is not None else ""
-
-            if delim is not None:
-                get_contents_to_file_using_delim()
-            else:
-                get_contents_to_file_oracle()
-        else:
-            # Write check for results if incomplete data available
-            # For folders
-            list = storage_conn.list_objects(namespace_name,bucket_name,prefix=object_name)
-            objects = list.data.objects
-            for one_path in objects:
-                name = one_path.name
-                if name.endswith('/') is False:
-                    log.info("Downloading file from %s" % name)
-                    response = storage_conn.get_object(namespace_name, bucket_name, name)
-                    content = response.data.content if response.data is not None else ""
-                    if delim is not None:
-                        get_contents_to_file_using_delim()
-                    else:
-                        get_contents_to_file_oracle()
-    except Exception as e:
-        raise Exception("Not able to download results: %s" % e.message)
-
-
-def _download_to_local_azure(blob_service, azure_path, fp, delim=None, skip_data_avail_check=False):
-    def get_contents_to_file_using_delim():
-        if sys.version_info < (3, 0, 0):
-            fp.write(str(content).replace(chr(1), delim))
-        else:
-            import io
-            if isinstance(fp, io.TextIOBase):
-                fp.buffer.write(content.decode('utf-8').replace(chr(1), delim).encode('utf8'))
-            elif isinstance(fp, io.BufferedIOBase) or isinstance(fp, io.RawIOBase):
-                fp.write(content.decode('utf8').replace(chr(1), delim).encode('utf8'))
-            else:
-                # Can this happen? Don't know what's the right thing to do in this case.
-                pass
-
-    def get_contents_to_file_azure():
-        if sys.version_info < (3, 0, 0):
-            fp.write(content.encode('utf8'))
-        else:
-            import io
-            if isinstance(fp, io.TextIOBase):
-                fp.buffer.write(content.encode('utf8'))
-            elif isinstance(fp, io.BufferedIOBase) or isinstance(fp, io.RawIOBase):
-                fp.write(content.encode('utf8'))
-            else:
-                # Can this happen? Don't know what's the right thing to do in this case.
-                pass
-
-
-    try:
-        m = Azure_URI_RE.match(azure_path)
-        container_name = m.group(1)
-        blob_name = m.group(3)
-        if azure_path.endswith('/') is False:
-
-            response = blob_service.get_blob_to_bytes(container_name, blob_name)
-            content = response.content
-
-            if delim is not None:
-                get_contents_to_file_using_delim()
-            else:
-                get_contents_to_file_azure()
-        else:
-            # Write check for results if incomplete data available
-            # For folders
-            list = blob_service.list_blobs(container_name, prefix=blob_name)
-            for blob in list:
-                name = blob.name
-                if name.endswith('/') is False:
-                    log.info("Downloading file from %s" % name)
-                    response = blob_service.get_blob_to_bytes(container_name, name)
-                    content = response.content
-                    if delim is not None:
-                        get_contents_to_file_using_delim()
-                    else:
-                        get_contents_to_file_azure()
-    except Exception as e:
-        raise Exception("Not able to download results: %s" % e.message)
